requirements.txt
pandas>=1.3.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# RAG Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0
pydantic>=2.0.0

# LLM Providers (choose one or both)
# OpenAI
openai>=1.0.0

# Google Gemini (Official Gen AI SDK)
google-genai>=1.31.0